refactor: addtional-info 페이지 리팩토링
